Move abstract list related RHs into module
Just like we have 'reglists.py' in the registration module. That way
'management.py' is much less bloaty
